+++ src/java/com/aboni/sensors/hw/BME280.java
@@ -144,9 +144,9 @@ public class BME280 implements Atmospheric {
                 int adcH = ((dataBuffer[6] & 0xff) << 8) + (dataBuffer[7] & 0xff);
 
                 // Temperature offset calculations
-                double var1 = (adcT / 16384.0 - digT[0] / 1024.0) * (double) digT[1];
+                double var1 = (adcT / 16384.0 - digT[0] / 1024.0) * digT[1];
                 double var2 = ((adcT / 131072.0 - digT[0] / 8192.0) *
-                        (adcT / 131072.0 - digT[0] / 8192.0)) * ((double) digT[2]);
+                        (adcT / 131072.0 - digT[0] / 8192.0)) * digT[2];
                 double tFine = (long) (var1 + var2);
                 cTemp = (var1 + var2) / 5120.0;
 
@@ -157,7 +157,7 @@ public class BME280 implements Atmospheric {
                 var2 = (var2 / 4.0) + (digP[3] * 65536.0);
                 var1 = (digP[2] * var1 * var1 / 524288.0 + digP[1] * var1) / 524288.0;
                 var1 = (1.0 + var1 / 32768.0) * digP[0];
-                double p = 1048576.0 - (double) adcP;
+                double p = 1048576.0 - adcP;
                 p = (p - (var2 / 4096.0)) * 6250.0 / var1;
                 var1 = digP[8] * p * p / 2147483648.0;
                 var2 = p * digP[7] / 32768.0;
